% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctionsThermo.R
\name{fileInfo.Thermo}
\alias{fileInfo.Thermo}
\title{fileInfo.Thermo}
\usage{
fileInfo.Thermo(filename, readIndex = TRUE, collapseCharacter = "-")
}
\arguments{
\item{filename}{name of the file from which the data is to be read. Must be a
Thermo Scientific mass spectrometry .raw file}

\item{readIndex}{logical vector. If TRUE the data element will contain info
on all scans in the file. If FALSE then data element will be empty}

\item{collapseCharacter}{some of the file info data consists of more than one
element. These will be pasted together where collapseCharacter defines the
separator}
}
\value{
a function that returns list of two objects: info (contains only
file info) and data (empty or index of scans)
}
\description{
function factory that generates a function for an info & data
element with info on a Thermo .raw mass spectrometry file
}
\examples{
demoRaw <- fs::path_package("extdata", "reserpine07.RAW", package = "readThermo")
result <- fileInfo.Thermo(filename = demoRaw)()
result[[1]]$info |> as.data.frame()
result[[1]]$data |> head(10)
result <- fileInfo.Thermo(filename = demoRaw, readIndex = FALSE)()
result[[1]]$info |> as.data.frame()
result[[1]]$data
}
