% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctionsThermo.R
\name{processingThermo.componentData}
\alias{processingThermo.componentData}
\title{processingThermo.componentData}
\usage{
processingThermo.componentData(sheet)
}
\arguments{
\item{sheet}{either a list or an element from a data.frame coming from
read.ThermoProcessing}
}
\value{
data.frame
}
\description{
extracts the component data from the data read by
read.ThermoProcessing
}
\examples{
demoXLS <- fs::path_package("extdata", "ExcelExp_Short.XLS", package = "readThermo")
result <- readProcessing.Thermo(filename = demoXLS)
processingThermo.componentData(result)
processingThermo.componentData(result[[1]])

}
