% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctionsThermo.R
\name{processingThermo.componentName}
\alias{processingThermo.componentName}
\title{processingThermo.componentName}
\usage{
processingThermo.componentName(sheet)
}
\arguments{
\item{sheet}{either a list or an element from a data.frame coming from
read.ThermoProcessing}
}
\value{
a character vector
}
\description{
extracts the component names from the data read by
read.ThermoProcessing
}
\examples{
demoXLS <- fs::path_package("extdata", "ExcelExp_Short.XLS", package = "readThermo")
result <- readProcessing.Thermo(filename = demoXLS)
processingThermo.componentName(result)
processingThermo.componentName(result[[1]])

}
