% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctionsThermo.R
\name{processingThermo.creatorInfo}
\alias{processingThermo.creatorInfo}
\title{processingThermo.creatorInfo}
\usage{
processingThermo.creatorInfo(sheet)
}
\arguments{
\item{sheet}{either a list or an element from a data.frame coming from
read.ThermoProcessing}
}
\value{
data.frame
}
\description{
extracts the creator (user) info from the data read by
read.ThermoProcessing
}
\examples{
demoXLS <- fs::path_package("extdata", "ExcelExp_Short.XLS", package = "readThermo")
result <- readProcessing.Thermo(filename = demoXLS)
processingThermo.creatorInfo(result)
processingThermo.creatorInfo(result[[1]])

}
