% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctionsThermo.R
\name{readChromatogram.Thermo}
\alias{readChromatogram.Thermo}
\title{readChromatogram.Thermo}
\usage{
readChromatogram.Thermo(
  filename,
  mz = NA,
  tolerance = 10,
  filter = "ms",
  type = "xic",
  additionalInfo = NA
)
}
\arguments{
\item{filename}{name of the .raw file from which the data is to be read}

\item{mz}{specifies the m/z('s) to make an extracted ion chormatogram of.
Ignored unless the 'type' argument is "xic"}

\item{tolerance}{spcifies the tolerance to use when extracting an ion
chromatogram. Ignored unless the 'type' argument is "xic". Please note that
this value is in 'ppm' and that a value of 10 is the same as a mass
tolerance of 5 ppm in eg the Thermo freestyle software (the 5 specified
there is a range m/z-5 till m/z+5). Here 10 means m/z-5 ppm till m/z+5 ppm.}

\item{filter}{specifies the scan filter to be used, default = "ms". Valid is
also eg "ms2" for ms2 data (if present in the file).}

\item{type}{specifies the data type to read, possible is: "tic"  (total ion
current), "bpc" (base peak chromatogram) or "xic" (extracted ion
chromatogram)}

\item{additionalInfo}{additional info to be added to the result of the read
data function. Should be named list format or a data.frame, default is NA}
}
\value{
a function that reads the data from the specified .raw file and returns a
list (of lists with two objects: info and data)
}
\description{
function factory that generates a function that reads data from
a thermo MS chromatogram file.
}
\note{
the file for this function needs to be '.raw' format. Internally the
'rawrr' package is used to do the actual data extraction
}
\examples{
demoRaw <- fs::path_package("extdata", "reserpine07.RAW", package = "readThermo")
result <- readChromatogram.Thermo(filename = demoRaw, type = "tic", filter = "ms2")()
result[[1]]$info
result[[1]]$data |> head(10)
with(result[[1]]$data, plot(rt, intensity, type = "l"))
demoRaw <- fs::path_package("extdata", "reserpine07.RAW", package = "readThermo")
result <- readChromatogram.Thermo(filename = demoRaw, type = "xic", filter = "ms2",
 mz = c(397.16, 448.13), tolerance = 500)()
result[[1]]$info
with(result[[1]]$data, plot(rt, intensity, type = "l"))
result[[2]]$info
with(result[[2]]$data, plot(rt, intensity, type = "l"))

}
